+++ app/models/guild-data.js
@@ -1,0 +1,8 @@
+const DiscordUtil = require("discordjs-util");
+
+module.exports = class GuildData {
+	constructor({ id, allowedRoles }) {
+		this.id = id;
+		this.allowedRoles = allowedRoles || [];
+	}
+};